src/storage/bigquery_loader.py
"""Loader para Google BigQuery"""

from google.cloud import bigquery
import logging

logger = logging.getLogger(__name__)


class BigQueryLoader:
    """Gerenciador de carregamento no BigQuery"""
    
    def __init__(self, config):
        """
        Inicializa cliente BigQuery
        
        Args:
            config: Dicionário com configurações
        """
        self.config = config
        # TODO: Inicializar cliente BigQuery
        pass
    
    def inserir_nfse_batch(self, nfse_list: list):
        """
        Insere lote de NFSe no BigQuery
        
        Args:
            nfse_list: Lista de dicionários com dados das NFSe
        """
        # TODO: Implementar insert batch
        logger.info(f"Inserindo {len(nfse_list)} NFSe no BigQuery")
        pass
