requirements.txt
# API SOAP
zeep==4.2.1
lxml==5.1.0
cryptography==41.0.7

# Google Cloud
google-cloud-bigquery==3.14.1
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4

# Utilities
PyYAML==6.0.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
